multi-mds/referent-inode: Migrate of referent inode on subtree migration
In a mulit-mds setup, let's say directories 'dir0', 'dir1' and 'dir2 are
created and pinned to mds.0, mds.1 and mds.2. Now a file say 'file1' is
created under 'dir0' and a harlink ('hl_file1') is created under 'dir1' for
'dir0/file1'. Now, since the hardlink is the first file created under
'dir1', the subtree migration kicks in after the link succeeds. So the
link operation is completed before subtree migration kicks in and the auth
would be mds.0.

The subtree migration should properly migrate the referent inode while
migration. It should remove the referent inode from the list of 'dirty_inodes'
which is journalled in 'mds.0' after migrating it to 'mds.1'. This
wasn't handled. This patch handles the migration part of the referent
inode.

Signed-off-by: Kotresh HR <khiremat@redhat.com>