mgr/cephadm: make enable_monitor_client configurable for nvmeof
Currently, the mon client work is not merged on main, but our
default nvmeof container will attempt to make use of it by default,
causing it to crash. This makes it configurable and defaults the
behavior to false. That can be changed once the work is actually
present in main.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 3276471ff765b0118105a7820a13d9e5efa9934b)