mgr/cephadm: don't mark daemons removed in the last minute as stray
There is sometimes a slight delay between when the core
mgr knows a daemon has been removed and when cephadm knows
it as been removed. This can cause stray daemon warnings
to pop up for a few seconds at a time. This patch tries
to avoid that by not marking daemons as stray that it
knows it just removed in the past minute.

Signed-off-by: Adam King <adking@redhat.com>