test/librados: Adjust AssertVersion tests
We write to the object second time to ensure that the
version number is greater than 0.
We should also ensure that the version number is greater than 1 because
otherwise passing (`v-1`) to `rados_write_op_assert_version` will result in
version of 0. As a result, -EINVAL would be returned (See case CEPH_OSD_OP_ASSERT_VER).

Luckily, since the classical OSD starts versioning from 1, the last version returned
after the second write is 2.
However, Crimson OSD starts versioning from 0 (and not from 1) so the test has to
be adjusted to also ensure that the version used is larger than 1.

* EINVAL case is also added to AssertVersion tests.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>