mgr/smb: add a new protocol for a FindingConfigStore
Add a new FindingConfigStore protocol. This FindingConfigStore is a
ConfigStore that provides a method for more quickly/efficiently finding
specific entries in the store. This is opposed to walking all keys and
getting all objects and comparing them. This protocol is optional and
will only be implemented by a few specialized stores.

Signed-off-by: John Mulligan <jmulligan@redhat.com>