mgr/smb: validate that a share name is unique per-cluster
We already require that the share_id be unique as it's part of the entry
key of the resource. Add a validation function to the handler that
checks if a (cluster_id, name) tuple is not duplicated. This function
uses the newly added config_store.find_in_store function to find the
potential dupes.

Signed-off-by: John Mulligan <jmulligan@redhat.com>