mgr/smb: add store transactions to parts of the handler
Add a wrapper contextmethod that initiates a transaction on a store if said
store supports transactions (otherwise the ctx manager is a no-op).
Use the wrapper over certain functions in the handler where many store
operations may be needed. This makes a big performance difference when
using a sqlite db backed store.

Signed-off-by: John Mulligan <jmulligan@redhat.com>