rgw: eliminate vault token perm for group read
Typically, the HashiCorp Vault token is refreshed by another process,
such as the Vault Agent. This process needs ownership of the file to
write the new token. Since Ceph primarily runs with its own user,
it cannot access the token file unless it has group read permissions.
This requires either assigning the correct group to the file by the
Vault Agent or ensuring Ceph is part of the appropriate group.

By eliminating the restriction on group read permissions, this can be achievable.

Fixes: https://tracker.ceph.com/issues/66121
Signed-off-by: Seena Fallah <seenafallah@gmail.com>