mgr/smb: Add acl_xattr configuration to share definition
Samba allows to store full Windows(NT) ACLs and bypass any lossy mapping
attempts to and from POSIX ACLs. Therefore we stack an additional module
in front of ceph to store NTACLs in a special extended attribute. Due to
certain limitations within containers the default xattr from security
namespace can't be used. Instead we configure a similar xattr from more
flexible 'user' namespace with the help of an option to the module.

ref: https://www.samba.org/samba/docs/current/man-html/vfs_acl_xattr.8.html
Signed-off-by: Anoop C S <anoopcs@cryptolab.net>