os/bluestore: Fix BlueRocksEnv attempts to use POSIX
How is it possible that we are mixing BlueFS and posix operations?

BlueRocksEnv implements rocksdb::Env interface.
To make things simpler it inherits from rocksdb::EnvWrapper.
The rocksdb::EnvWrapper is initialized from DefaultEnv, which brings
default implementation for threads and files.
We override file-related interface.

The problem is when we forget to implement something, or if rocksdb::Env
interface gets expanded. The function returns ENOTSUPP, and rocksdb can
handle it. But it does not matter, because it is implemented in DefaultEnv.

In future we should drop inheriting from EnvWrapper and call DefaultEnv
directly wherever we need.

Fixes https://tracker.ceph.com/issues/66717.

Signed-off-by: Adam Kupczyk <akupczyk@ibm.com>