multi-mds: Notify replica on unlink of hardlink
Problem:
The unlink of the hardlink asserts at "linkage_is_null()" in
decode_lock_state called by handle_simple_lock

Analysis:
This is happening because of loading all referent inodes into memory. This results
in CDentry of the secondary link files to be replicated in all the mdses.
The unlink received in one mds needs to inform the replicated mds about this
and linkage would be made null which is not happening. When the 'drop_locks'
is being called during response request, the source mds has linkage null but
the replicated mds has linkage non null.

Signed-off-by: Kotresh HR <khiremat@redhat.com>