multi-mds/linkmerge: Fix send_dentry_link
The 'send_dentry_link' is common api used to notify
replicas for all the inode-dentry modifying operations.
But this was ignoring rename witnesses as required
changes are done to witnesses at the very beginning
of rename op. But the linkmerge, an internal rename
because of stray re-integration is an exception
where the rename witnesses needs to be notified of
the referent inode list change in the real inode.
There is no other case in rename where the real
inode fields change and hence it wasn't notified after
rename.

Signed-off-by: Kotresh HR <khiremat@redhat.com>