mgr/rgw: add ability for rgw realm bootstrap to create pools
Specifically by setting the data_pool_attributes field in
the RGW spec. For example

service_type: rgw
service_id: my_realm.my_zone
placement:
  count: 2
spec:
  data_pool_attributes:
    type: ec
    k: 6
    m: 2
  rgw_realm: my_realm
  rgw_zone: my_zone
  rgw_zonegroup: my_zonegroup

would create the my_zone.rgw.buckets.data pool as an
ec pool with a new ec porfile with k=8 and m=2. Currently
the ec profile fields this spec will make use of
are only "k", "m", "pg_num" and "crush-device-class".
If other attributes are set, or if the pool type is
"replicated" the key value pairs are assumed to
be ones that can be passed to the "ceph osd pool create"
command. The other pools for the rgw zone (e.g. buckets
index pool) are all made as replicated pool with
defaults settings

Signed-off-by: Adam King <adking@redhat.com>