qa: fail mgr between tests to start with clean state
Problem 1: The mgr declares itself as available as soon as it dispatches
code to the finisher to load the python extension modules. This causes the
mgr dump command to return the mgr as available even when the modules aren't
in the active state to handle commands.

Solution 1: Sleep for a minute to allow the mgr finisher to instantiate the
python extension modules so that they can start handling commands.
-----
Problem 2: Mgr uses stale mounts to fetch clone jobs.

Solution 2: Fail the mgr between two tests so that there is a failover and
a state cleanup.

Fixes: https://tracker.ceph.com/issues/66009
Signed-off-by: Milind Changire <mchangir@redhat.com>