qa: move "rbd:nvmeof" sub-suite outside rbd
Make nvmeof suite as an independent suite (outside of rbd).
This is because nvmeofgwmon is disabled by default in main.
So the nvmeof teuthology test would only be tested where
the nvmeofgwmon is enabled.
In future, it can be symlinked with rbd suite.

Also fix cephadm task's "watchdog_setup" method.

Signed-off-by: Vallari Agrawal <val.agl002@gmail.com>