erasure-code: Implement minimal reads - message send
Adapt the functions which determine the extents of each shard that
need to be read, such that they pick the minimal reads possible.

This code was tested with subsequent test harness commits only and is
not expected to work outside such an environment, as the IO complete
path cannot yet cope with minimal reads.

There is quite a bit of fiddly integer maths going on here and this
will be neatened up in a later commit once further testing has taken
place and this maths is no longer undergoing frequent changes.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>