include/ceph_features: add NVMEOFHA feature bit
Normally, we'd just use the SERVER_SQUID or SERVER_T flags instead of
using an extra.  However, the nvmeof ha monitor paxos service has had a
more complex development journey.  There are users interested in using
the nvmeof ha feature in squid, but it didn't make the cutoff for
backporting it.  There's an upstream nvmeof-squid branch in the ceph.git
repository with the patches backported for anyone interested in building
it.

However, that means that users of our normal stable releases will see
the feature added to the monitor one release after anyone who chooses to
use the nvmeof-squid branch.  We could disallow upgrades from
nvmeof-squid to T, but by adding a feature bit here we make such a
restriction unnecessary.

Signed-off-by: Samuel Just <sjust@redhat.com>