rgw: send self zonegroup on forward to master
When creating a bucket in the secondary zonegroup, the `rgwx-zonegroup`
is mistakenly set to the master zonegroup when forwarding the request.
Instead, it should be set to the secondary zonegroup so that the master
zonegroup creates the bucket for the intended zonegroup rather than
for itself.

Fixes: https://tracker.ceph.com/issues/67610
Signed-off-by: Seena Fallah <seenafallah@gmail.com>
(cherry picked from commit 9769fb2656b723c5708b51935829aa7ac77e1d63)