cephadm: move envs_list logic back to base class
When the ContainerCommon base class got created it added a very simple
stub `envs_list` method. But there should be no need for a fully
implemented envs_list not to exist on the base class because the method
exits solely to convert between the dict-returning `envs` method and
list-of-strings expecting code elsewhere in cephadm. I missed this in
review and so am just fixing it up here.

Signed-off-by: John Mulligan <jmulligan@redhat.com>