mgr; common/options: reorder checks & add retry logic for module loading
The current "active and enabled" check order can result in an ENOTSUP being
returned when a module is enabled, but needs more time to load. In this case,
the command should be supported, but the module timed out loading. With the internal
retry logic that is added (the mgr retries loading the module several
times before giving up), ETIMEDOUT is more appropriate.

The default max time that a module can take to load is now 5 seconds. To
adjust this time, the following command can be run:
  `ceph config set mgr mgr_module_load_timeout <uint>`

The default cadence at which the mgr retries is every 1 second. To
adjust this time, the following command can be run:
  `ceph config set mgr mgr_module_load_interval <uint>`

Fixes: https://tracker.ceph.com/issues/69012
Signed-off-by: Laura Flores <lflores@ibm.com>
Co-authored-by: Brad Hubbard <bhubbard@redhat.com>