osd: EC partial metadata writes - bug fix
Partial metadata writes sometimes performed incorrect
log recovery by choosing a shard that had not participated in
writes as the authorative log and then rolling back "diverged"
writes that had actually been committed and completed to the
client.

Fix is to use the partial_write_last_complete state to
move the head of the log for the shard providing the
authorative log forward to include committed writes.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>