mds: Fix struct::remotebit decode
The struct::remotebit::decode wasn't decoding the
referent inode number and referent inode structure
when reading the journal event from the log.
But those did got encoded and written to the journal log.

The boolean referent variable maintained in struct::remotebit
is incorrectly used in the decode.

This was causing the following issues.
1. The mds journal(MDLog) replay (EMetaBlob::replay) of hardlink
   was failing.
2. The cephfs-journal-tool's recover_dentries sub command fails
   to recover hardlinked dentry from the journal

TODO: Add test for the above

Signed-off-by: Kotresh HR <khiremat@redhat.com>