qa/cephadm: refresh daemon state during test_pause test
The previous test ends with `ceph orch daemon restart osd.0`
which may temporarily cause a CEPHADM_FAILED_DAEMON warning
for that OSD. In my testing with interactive reruns this
either didn't happen or, if it did, the warning would clear
itself up a bit later as soon as cephadm went and refreshed
the daemon state on the host. This commit makes cephadm
refresh the daemon state after resuming the orchestrator
so that temporary artifacts from the last test don't cause
this one to fail.

Fixes: https://tracker.ceph.com/issues/69526

Signed-off-by: Adam King <adking@redhat.com>